Add Logic to Prevent Final Encounters From Being Submitted Without Previous Encounters.

+++ application/models/QuickSubmissionModel/class/QuickSubmissionModel.php
@@ -4,9 +4,10 @@
  * submit kills without logging in page
  */
 class QuickSubmissionModel extends Model {
-    protected $_validSubmission = false;
-    protected $_validEncounter = false;
-    protected $_validScreenshot = false;
+    protected $_validSubmission      = false;
+    protected $_validEncounter       = false;
+    protected $_validScreenshot      = false;
+    protected $_finalEncounterExists = false;
     protected $_guildId;
     protected $_encounterId;
     protected $_guildDetails;
@@ -27,19 +28,25 @@ class QuickSubmissionModel extends Model {
 
         if ( Post::formActive() ) { // Form has required fields filled out
             $this->_validSubmission = $this->validateForm();
-
+            
             if ( $this->_validSubmission ) { // Ensures guild does not have encounter already submitted
                 $this->_encounterExists = $this->validateEncounter();
 
                 if ( $this->_encounterExists ) { // Ensures a valid screenshot is submitted
-                    $this->_validScreenshot = Functions::validateImage($this->_formFields->screenshot);
+                    $this->_finalEncounterExists = $this->validateFinalEncounter();
 
-                    if ( $this->_validScreenshot ) { // Submit the data
-                        $this->processForm();
+                    if ( $this->_finalEncounterExists ) { // Ensures guild already has final encounter submitted
+                        $this->_validScreenshot = Functions::validateImage($this->_formFields->screenshot);
 
-                        $this->_dialogOptions = array('title' => 'Success', 'message' => 'Your kill has been submitted successfully! Standings and Rankings will be updated accordingly!');
+                        if ( $this->_validScreenshot ) { // Submit the data
+                            $this->processForm();
+
+                            $this->_dialogOptions = array('title' => 'Success', 'message' => 'Your kill has been submitted successfully! Standings and Rankings will be updated accordingly!');
+                        } else {
+                            $this->_dialogOptions = array('title' => 'Error', 'message' => 'All kill submissions require a valid screenshot. Please submit your kill with one!');
+                        }
                     } else {
-                        $this->_dialogOptions = array('title' => 'Error', 'message' => 'All kill submissions require a valid screenshot. Please submit your kill with one!');
+                        $this->_dialogOptions = array('title' => 'Error', 'message' => 'Please submit all encounters before the final encounter!');
                     }
                 } else {
                     $this->_dialogOptions = array('title' => 'Error', 'message' => 'Please choose a valid encounter! (The encounter you selected may have already been submitted)');
@@ -93,6 +100,23 @@ class QuickSubmissionModel extends Model {
      */
     public function validateEncounter() {
         return !isset($this->_guildDetails->_encounterDetails->{$this->_encounterId});
+    }
+
+    /**
+     * validate if final encounter exists
+     * 
+     * @return boolean [ true if final encounter exists]
+     */
+    public function validateFinalEncounter() {
+        $encounterDetails = CommonDataContainer::$encounterArray[$this->_encounterId];
+        $dungeonDetails   = CommonDataContainer::$dungeonArray[$encounterDetails->_dungeonId];
+        $finalEncounterId = $dungeonDetails->_finalEncounterId;
+
+        if ( isset($this->_guildDetails->_encounterDetails->finalEncounterId) ) {
+            $this->_finalEncounterExists = true;
+        }
+
+        return $this->_finalEncounterExists;
     }
 
     /**
